+++ lib/src/styles/theme.dart
@@ -1,5 +1,4 @@
 import 'package:fluent_ui/fluent_ui.dart';
-
 import 'package:flutter/foundation.dart';
 import 'package:flutter/rendering.dart';
 
@@ -198,19 +197,15 @@ class FluentThemeData with Diagnosticable {
 
   final NavigationPaneThemeData navigationPaneTheme;
   final BottomNavigationThemeData bottomNavigationTheme;
-  final BottomSheetThemeData bottomSheetTheme;
   final CheckboxThemeData checkboxTheme;
-  final ChipThemeData chipTheme;
   final ContentDialogThemeData dialogTheme;
   final DividerThemeData dividerTheme;
   final FocusThemeData focusTheme;
   final IconThemeData iconTheme;
   final InfoBarThemeData infoBarTheme;
-  final PillButtonBarThemeData pillButtonBarTheme;
   final RadioButtonThemeData radioButtonTheme;
   final ScrollbarThemeData scrollbarTheme;
   final SliderThemeData sliderTheme;
-  final SnackbarThemeData snackbarTheme;
   final ToggleButtonThemeData toggleButtonTheme;
   final ToggleSwitchThemeData toggleSwitchTheme;
   final TooltipThemeData tooltipTheme;
@@ -242,10 +237,8 @@ class FluentThemeData with Diagnosticable {
     Duration? slowAnimationDuration,
     Curve? animationCurve,
     BottomNavigationThemeData? bottomNavigationTheme,
-    BottomSheetThemeData? bottomSheetTheme,
     ButtonThemeData? buttonTheme,
     CheckboxThemeData? checkboxTheme,
-    ChipThemeData? chipTheme,
     ToggleSwitchThemeData? toggleSwitchTheme,
     IconThemeData? iconTheme,
     ContentDialogThemeData? dialogTheme,
@@ -256,10 +249,8 @@ class FluentThemeData with Diagnosticable {
     ToggleButtonThemeData? toggleButtonTheme,
     SliderThemeData? sliderTheme,
     InfoBarThemeData? infoBarTheme,
-    PillButtonBarThemeData? pillButtonBarTheme,
     FocusThemeData? focusTheme,
     ScrollbarThemeData? scrollbarTheme,
-    SnackbarThemeData? snackbarTheme,
     ResourceDictionary? resources,
   }) {
     brightness ??= Brightness.light;
@@ -293,7 +284,6 @@ class FluentThemeData with Diagnosticable {
     focusTheme ??= const FocusThemeData();
     buttonTheme ??= const ButtonThemeData();
     checkboxTheme ??= const CheckboxThemeData();
-    chipTheme ??= const ChipThemeData();
     toggleButtonTheme ??= const ToggleButtonThemeData();
     toggleSwitchTheme ??= const ToggleSwitchThemeData();
     iconTheme ??= isLight
@@ -314,11 +304,8 @@ class FluentThemeData with Diagnosticable {
     radioButtonTheme ??= const RadioButtonThemeData();
     sliderTheme ??= const SliderThemeData();
     infoBarTheme ??= const InfoBarThemeData();
-    pillButtonBarTheme ??= const PillButtonBarThemeData();
     scrollbarTheme ??= const ScrollbarThemeData();
     bottomNavigationTheme ??= const BottomNavigationThemeData();
-    snackbarTheme ??= const SnackbarThemeData();
-    bottomSheetTheme ??= const BottomSheetThemeData();
 
     return FluentThemeData.raw(
       brightness: brightness,
@@ -340,7 +327,6 @@ class FluentThemeData with Diagnosticable {
       bottomNavigationTheme: bottomNavigationTheme,
       buttonTheme: buttonTheme,
       checkboxTheme: checkboxTheme,
-      chipTheme: chipTheme,
       dialogTheme: dialogTheme,
       dividerTheme: dividerTheme,
       focusTheme: focusTheme,
@@ -354,9 +340,6 @@ class FluentThemeData with Diagnosticable {
       toggleSwitchTheme: toggleSwitchTheme,
       tooltipTheme: tooltipTheme,
       typography: typography,
-      snackbarTheme: snackbarTheme,
-      pillButtonBarTheme: pillButtonBarTheme,
-      bottomSheetTheme: bottomSheetTheme,
       menuColor: menuColor,
       cardColor: cardColor,
       resources: resources,
@@ -384,7 +367,6 @@ class FluentThemeData with Diagnosticable {
     required this.micaBackgroundColor,
     required this.buttonTheme,
     required this.checkboxTheme,
-    required this.chipTheme,
     required this.toggleSwitchTheme,
     required this.bottomNavigationTheme,
     required this.iconTheme,
@@ -398,9 +380,6 @@ class FluentThemeData with Diagnosticable {
     required this.infoBarTheme,
     required this.focusTheme,
     required this.scrollbarTheme,
-    required this.snackbarTheme,
-    required this.pillButtonBarTheme,
-    required this.bottomSheetTheme,
     required this.menuColor,
     required this.cardColor,
     required this.resources,
@@ -440,7 +419,6 @@ class FluentThemeData with Diagnosticable {
       animationCurve: t < 0.5 ? a.animationCurve : b.animationCurve,
       buttonTheme: ButtonThemeData.lerp(a.buttonTheme, b.buttonTheme, t),
       checkboxTheme: CheckboxThemeData.lerp(a.checkboxTheme, b.checkboxTheme, t),
-      chipTheme: ChipThemeData.lerp(a.chipTheme, b.chipTheme, t),
       toggleSwitchTheme: ToggleSwitchThemeData.lerp(a.toggleSwitchTheme, b.toggleSwitchTheme, t),
       iconTheme: IconThemeData.lerp(a.iconTheme, b.iconTheme, t),
       dialogTheme: ContentDialogThemeData.lerp(a.dialogTheme, b.dialogTheme, t),
@@ -456,10 +434,6 @@ class FluentThemeData with Diagnosticable {
       scrollbarTheme: ScrollbarThemeData.lerp(a.scrollbarTheme, b.scrollbarTheme, t),
       bottomNavigationTheme:
           BottomNavigationThemeData.lerp(a.bottomNavigationTheme, b.bottomNavigationTheme, t),
-      snackbarTheme: SnackbarThemeData.lerp(a.snackbarTheme, b.snackbarTheme, t),
-      pillButtonBarTheme:
-          PillButtonBarThemeData.lerp(a.pillButtonBarTheme, b.pillButtonBarTheme, t),
-      bottomSheetTheme: BottomSheetThemeData.lerp(a.bottomSheetTheme, b.bottomSheetTheme, t),
       menuColor: Color.lerp(a.menuColor, b.menuColor, t)!,
       selectionColor: Color.lerp(a.selectionColor, b.selectionColor, t)!,
     );
@@ -506,9 +480,7 @@ class FluentThemeData with Diagnosticable {
     Curve? animationCurve,
     ButtonThemeData? buttonTheme,
     BottomNavigationThemeData? bottomNavigationTheme,
-    BottomSheetThemeData? bottomSheetTheme,
     CheckboxThemeData? checkboxTheme,
-    ChipThemeData? chipTheme,
     ToggleSwitchThemeData? toggleSwitchTheme,
     IconThemeData? iconTheme,
     ContentDialogThemeData? dialogTheme,
@@ -519,10 +491,8 @@ class FluentThemeData with Diagnosticable {
     ToggleButtonThemeData? toggleButtonTheme,
     SliderThemeData? sliderTheme,
     InfoBarThemeData? infoBarTheme,
-    PillButtonBarThemeData? pillButtonBarTheme,
     FocusThemeData? focusTheme,
     ScrollbarThemeData? scrollbarTheme,
-    SnackbarThemeData? snackbarTheme,
     ResourceDictionary? resources,
   }) {
     return FluentThemeData.raw(
@@ -548,15 +518,12 @@ class FluentThemeData with Diagnosticable {
       animationCurve: animationCurve ?? this.animationCurve,
       buttonTheme: this.buttonTheme.merge(buttonTheme),
       bottomNavigationTheme: this.bottomNavigationTheme.merge(bottomNavigationTheme),
-      bottomSheetTheme: this.bottomSheetTheme.merge(bottomSheetTheme),
       checkboxTheme: this.checkboxTheme.merge(checkboxTheme),
-      chipTheme: this.chipTheme.merge(chipTheme),
       dialogTheme: this.dialogTheme.merge(dialogTheme),
       dividerTheme: this.dividerTheme.merge(dividerTheme),
       focusTheme: this.focusTheme.merge(focusTheme),
       iconTheme: this.iconTheme.merge(iconTheme),
       infoBarTheme: this.infoBarTheme.merge(infoBarTheme),
-      pillButtonBarTheme: this.pillButtonBarTheme.merge(pillButtonBarTheme),
       navigationPaneTheme: this.navigationPaneTheme.merge(navigationPaneTheme),
       radioButtonTheme: this.radioButtonTheme.merge(radioButtonTheme),
       scrollbarTheme: this.scrollbarTheme.merge(scrollbarTheme),
@@ -564,7 +531,6 @@ class FluentThemeData with Diagnosticable {
       toggleButtonTheme: this.toggleButtonTheme.merge(toggleButtonTheme),
       toggleSwitchTheme: this.toggleSwitchTheme.merge(toggleSwitchTheme),
       tooltipTheme: this.tooltipTheme.merge(tooltipTheme),
-      snackbarTheme: this.snackbarTheme.merge(snackbarTheme),
       resources: resources ?? this.resources,
     );
   }
@@ -593,16 +559,8 @@ class FluentThemeData with Diagnosticable {
         'mediumAnimationDuration',
         mediumAnimationDuration,
       ))
-      ..add(DiagnosticsProperty<Duration>(
-        'fastAnimationDuration',
-        fastAnimationDuration,
-      ))
-      ..add(DiagnosticsProperty<Duration>(
-        'fasterAnimationDuration',
-        fasterAnimationDuration,
-      ))
-      ..add(
-        DiagnosticsProperty<Curve>('animationCurve', animationCurve),
-      );
+      ..add(DiagnosticsProperty<Duration>('fastAnimationDuration', fastAnimationDuration))
+      ..add(DiagnosticsProperty<Duration>('fasterAnimationDuration', fasterAnimationDuration))
+      ..add(DiagnosticsProperty<Curve>('animationCurve', animationCurve));
   }
 }